requirements.txt
openai
hyperon
uagents
uagents-core
python-dotenv
requests
